+++ models/analyticsVideo.py
@@ -1,4 +1,4 @@
-from pydantic import BaseModel
+from pydantic import BaseModel, Field
 from typing import Dict, List, Optional
 from datetime import datetime
 
@@ -27,6 +27,4 @@ class VideoAnalytics(BaseModel):
 
     class Config:
         # Allow population of fields with alias names
-        fields = {
-            "id": "_id"  
-        }
+        my_field: str = Field(alias="myField")
