+++ models/review.py
@@ -1,13 +1,6 @@
-from pydantic import BaseModel
-from typing import List, Optional
+from pydantic import BaseModel,Field
+from typing import Optional
 from datetime import datetime
-
-class RatingDistribution(BaseModel):
-    five_star: str
-    four_star: str
-    three_star: str
-    two_star: str
-    one_star: str
 
 class RecentReview(BaseModel):
     product_id: str
@@ -15,17 +8,10 @@ class RecentReview(BaseModel):
     rating: float
     title: str
     comment: Optional[str]
-    date: str
     verified_purchase: bool
     created_at: Optional[str] = datetime.now().isoformat()
     updated_at: Optional[str] = datetime.now().isoformat()
 
     class Config:
         # Allow population of fields with alias names
-        fields = {
-            "id": "_id"  
-        }
-
-class ReviewsResponse(BaseModel):
-    status: str
-    reviews: List[RecentReview]
+        my_field: str = Field(alias="myField")
