+++ models/video.py
@@ -1,8 +1,9 @@
-from pydantic import BaseModel
+from pydantic import BaseModel, Field
 from typing import List, Optional
 from datetime import datetime
 
 class Video(BaseModel):
+    id: str
     title: str
     category: str
     subcategory: str
@@ -17,6 +18,4 @@ class Video(BaseModel):
 
     class Config:
         # Allow population of fields with alias names
-        fields = {
-            "id": "_id"  
-        }
+        my_field: str = Field(alias="myField")
